requirements.txt
Django
djangorestframework
django-cors-headers
environs